SessionManager: use unsigned int for animation name pair

+++ src/SessionManager.hpp
@@ -35,7 +35,7 @@ public:
         struct CellanimData {
             std::shared_ptr<RvlCellAnim::RvlCellAnimObject> object;
             std::string name;
-            std::unordered_map<uint16_t, std::string> animNames;
+            std::unordered_map<unsigned, std::string> animNames;
         };
         std::vector<CellanimData> cellanims;
         std::vector<std::shared_ptr<Texture>> sheets;
@@ -61,7 +61,7 @@ public:
         std::shared_ptr<Texture>& getCellanimSheet() {
             return this->sheets.at(std::min<unsigned>(this->getCellanimObject()->sheetIndex, this->sheets.size() - 1));
         }
-        std::unordered_map<uint16_t, std::string>& getAnimationNames() {
+        std::unordered_map<unsigned, std::string>& getAnimationNames() {
             return this->cellanims.at(this->currentCellanim).animNames;
         }
