+++ src/command/CommandSwapAnimations.hpp
@@ -98,7 +98,7 @@ private:
             ->animations;
     }
 
-    std::unordered_map<uint16_t, std::string>& getAnimationNames() {
+    std::unordered_map<unsigned, std::string>& getAnimationNames() {
         return
             SessionManager::getInstance().getCurrentSession()
             ->cellanims.at(this->cellanimIndex).
